// demo/src/app/api/administration/models/fleet-fleet-_set-_address-_contact.ts
/* tslint:disable */
import { ContactFleet_set_address_contact } from './contact-fleet-_set-_address-_contact';

/**
 * A Fleet
 */
export interface FleetFleet_set_address_contact {
  name: string;

  /**
   * The Description of the fleet
   */
  description?: string;
  contact?: ContactFleet_set_address_contact;
}